Refactor subject analysis around stimulus types
